clean up code, get plot working. color use still need work
