+++ mapplot.py
@@ -1,7 +1,7 @@
 from collections import deque
-trail = deque()
+pastpoints = deque()
 
-def plot(colorArray, pointArray):
+def plot(colorArray, pointArray, isstrail=False):
   rows = 30
   cols = 40
   rowList = [76.5,71.4,66.3,61.2,56.1,51,45.9,40.8,35.7,30.6,25.5,20.4,15.3,10.2,5.1,0,-5.1,-10.2,-15.3,-20.4,-25.5,-30.6,-35.7,-40.8,-45.9,-51,-56.1,-61.2,-66.3,-71.4,-76.5]
@@ -15,33 +15,28 @@ def plot(colorArray, pointArray):
   lightArray = []
   for k in colorArray:
     lightArray.append(k[0])
-
-  print(lightArray)
   
   string = ""
   for i in rowList:
     # string = string + str(rowList.index(i)).ljust(2, ' ') + " |" 
     string = string + str(i).ljust(5, ' ') + " |" 
     for j in colList:
-      light = False
-      old = False
-      map = False
+      light = trail = map = False
       # print(pointArray.index((i,j) in lightArray)
       # print((j,i))
       if ((i,j) in pointArray):
         if(pointArray.index((i,j)) in lightArray):
           light = True
-        elif(pointArray.index((i,j)) in trail):
-          old = True
+        elif(isstrail and pointArray.index((i,j)) in pastpoints):
+          trail = True
         else:
           map = True
       
-
       if i==0:
         string= string + "  -  "
       elif(light):
         string = string + "  O  "
-      elif(old):
+      elif(trail):
         string = string + "  +  "
       elif(map):
         string = string + "  .  "
@@ -52,14 +47,14 @@ def plot(colorArray, pointArray):
   string = string + ">"
   print("\r" + string, end = ">")
 
+  if (isstrail):
     #this only works if lightray is single int
-  if len(lightArray) == 1:
-    led = lightArray.pop()
-    if led not in trail:
-      trail.append(led)
-  while len(trail)>20:
-    trail.popleft()
-  print(trail)
+    if len(lightArray) == 1:
+      led = lightArray.pop()
+      if led not in pastpoints:
+        pastpoints.append(led)
+    while len(pastpoints)>20:
+      pastpoints.popleft()
 
 # if __name__ == '__main__':
